Fix connection failures by destroying connections rather than releasing.
This seems uneccessary to me however after the connection was released
it was still managing to time out.

Destroying after use will not be much different as the pool will create
a new connection when required.

The error in being sent through the response has also been taken out as
the client does not need to know of this error.